Rename .nft to .nfts and remove "getAll"
