+++ packages/libs/sdk/spec/api/client.ts
@@ -1,7 +1,7 @@
-import { Api } from '../../src';
+import { API } from '../../src';
 
 const opts: Record<string, any> = {
   gqlApiKey: process.env['QUICKNODE_GQL_API_KEY'] || '',
 };
 
-export const apiClient = new Api(opts);
+export const apiClient = new API(opts);
