+++ packages/libs/sdk/src/index.ts
@@ -1,5 +1,5 @@
 import QuickNode from './client';
-import { Api } from './api';
+import { API } from './api';
 
 export default QuickNode;
-export { Api };
+export { API };
